fix 46M memmap encode OOM: chunked encode_df_to_memmap (bounded ~1GB peak vs ~150GB one-shot; SIGKILL loop flows 4299/4300) + drop packbits astype copy + revive stale-import encoder tests

+++ src/wnn/representations/thermometer.py
@@ -204,7 +204,9 @@ class ThermometerEncoder:
 		total_bits = int(bool_matrix.shape[1])
 		# np.packbits with bitorder='little' matches Rust PackedBits layout
 		# (LSB-first within byte). Each row packs to ceil(total_bits/8) bytes.
-		packed = np.packbits(bool_matrix.astype(np.uint8), axis=1, bitorder='little')
+		# packbits accepts bool natively — an astype(np.uint8) here would be
+		# a second full-matrix copy on top of the hstack.
+		packed = np.packbits(bool_matrix, axis=1, bitorder='little')
 		return packed, total_bits
 
 	def iter_chunks(self, df, chunk_size: int):
